Fixed more warnings (comparizon between signed / unsigned values)

+++ beziercurves/drawable_lines.cpp
@@ -55,7 +55,7 @@ void DrawableLines::DefinePairsGeometry(std::vector<Point*>* points)
 	}
 
 	// Generate the VAO (if it hasn't yet)
-	if (_vao == -1)
+	if ((int)_vao == -1)
 	{
 		glGenVertexArrays(1, &_vao);
 	}
@@ -64,7 +64,7 @@ void DrawableLines::DefinePairsGeometry(std::vector<Point*>* points)
 	glBindVertexArray(_vao);
 
 	// Generate the VBO (if it hasn't yet)
-	if (_vbo == -1)
+	if ((int)_vbo == -1)
 	{
 		glGenBuffers(1, &_vbo);
 	}
@@ -74,7 +74,7 @@ void DrawableLines::DefinePairsGeometry(std::vector<Point*>* points)
 	glBufferData(GL_ARRAY_BUFFER, nPoints * 3 * sizeof(GLfloat), vertices, GL_STATIC_DRAW);
 
 	// Generate the EBO (if it hasn't yet)
-	if (_ebo == -1)
+	if ((int)_ebo == -1)
 	{
 		glGenBuffers(1, &_ebo);
 	}
@@ -142,7 +142,7 @@ void DrawableLines::DefineSequenceGeometry(std::vector<Point*>* points)
 	}
 
 	// Generate the VAO (if it hasn't yet)
-	if (_vao == -1)
+	if ((int)_vao == -1)
 	{
 		glGenVertexArrays(1, &_vao);
 	}
@@ -151,7 +151,7 @@ void DrawableLines::DefineSequenceGeometry(std::vector<Point*>* points)
 	glBindVertexArray(_vao);
 
 	// Generate the VBO (if it hasn't yet)
-	if (_vbo == -1)
+	if ((int)_vbo == -1)
 	{
 		glGenBuffers(1, &_vbo);
 	}
@@ -161,7 +161,7 @@ void DrawableLines::DefineSequenceGeometry(std::vector<Point*>* points)
 	glBufferData(GL_ARRAY_BUFFER, nPoints * 3 * sizeof(GLfloat), vertices, GL_STATIC_DRAW);
 
 	// Generate the EBO (if it hasn't yet)
-	if (_ebo == -1)
+	if ((int)_ebo == -1)
 	{
 		glGenBuffers(1, &_ebo);
 	}
